Make ByteConsumer own its reader
This is in preparation for making pq multi-threaded; the trait `Read`
is not `Send` and so with the previous declaration (read: &'a mut
Read) ByteConsumer was _not_ Send and could not be moved into a
closure, i.e., with thread::spawn.

This commit modifies the API of ByteConsumer by making it take
ownership of a value of type T where T implements the Read trait.