Make decode part of decode_or_convert concurrent
This commit adds concurrency to the decoding part of
CommandRunner::decode_or_convert.  A pair of threads -- a producer and
a consumer -- are spawned; the producer decodes Protobuf values and
sends the decoded Value over a channel to the consumer who encodes the
values to JSON and outputs them to stdout.