Switch kafka provider to rdkafka
It appears the rust kafka implementation was subtly broken.
(I did not get any messages but pq exited immediately)

This commit changes the provider for the kafka stream to rdkafka (a
wrapper of the official rdkafka library).

There should be no outwardly visible api changes, aside from the type of
the stream-delimit Error.